[search] Moved out common test code to search_tests_support.
