+++ search/search_integration_tests/search_integration_tests.pro
@@ -6,7 +6,7 @@ CONFIG -= app_bundle
 TEMPLATE = app
 
 ROOT_DIR = ../..
-DEPENDENCIES = generator_tests_support generator routing search storage stats_client indexer \
+DEPENDENCIES = generator_tests_support generator routing search search_tests_support storage stats_client indexer \
                platform geometry coding base tess2 protobuf tomcrypt jansson
 
 DEPENDENCIES += opening_hours \
@@ -22,9 +22,5 @@ SOURCES += \
     ../../testing/testingmain.cpp \
     retrieval_test.cpp \
     smoke_test.cpp \
-    test_search_engine.cpp \
-    test_search_request.cpp \
 
 HEADERS += \
-    test_search_engine.hpp \
-    test_search_request.hpp \
