+++ search/search_tests_support/test_search_request.hpp
@@ -9,6 +9,10 @@
 #include "std/string.hpp"
 #include "std/vector.hpp"
 
+namespace search
+{
+namespace tests_support
+{
 class TestSearchEngine;
 
 // This class wraps a search query to a search engine. Note that the
@@ -33,3 +37,5 @@ private:
   vector<search::Result> m_results;
   bool m_done;
 };
+}  // namespace tests_support
+}  // namespace search
